"Refactor: DTO 경로 변경 및 엔티티 내부 논리 삭제 메소드명 변경"
